mcp-kroger/pyproject.toml
[project]
name = "mcp-kroger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kroger-api==0.2.0",
    "python-dotenv>=1.0",
    "typer>=0.12",
]

[project.scripts]
kroger = "kroger_cli:app"

[tool.setuptools.packages.find]
where = ["src"]